fix new development

+++ app/src/main/java/com/pbt/myfarm/Activity/task_object/ViewTaskObjectViewModel.kt
@@ -1,25 +1,26 @@
 package com.pbt.myfarm.Activity.task_object
 
-import android.app.Activity
+
+import android.app.AlertDialog
 import android.app.Application
-import android.app.Dialog
 import android.content.Context
+import android.content.DialogInterface
+import android.content.Intent
 import android.util.Log
-import android.view.Window
+import android.widget.Toast
 import androidx.lifecycle.AndroidViewModel
 import androidx.lifecycle.MutableLiveData
 import com.google.gson.Gson
 import com.pbt.myfarm.Activity.TaskFunctions.ListTaskFunctions
+import com.pbt.myfarm.Activity.ViewTask.ViewTaskActivity
 import com.pbt.myfarm.BaseHttpResponse
+import com.pbt.myfarm.DataBase.DbHelper
 import com.pbt.myfarm.HttpResponse.BaseTaskFunction
 import com.pbt.myfarm.HttpResponse.HttpResponse
-import com.pbt.myfarm.OffLineSyncModel
-import com.pbt.myfarm.R
 import com.pbt.myfarm.Service.ApiClient
 import com.pbt.myfarm.Service.ApiInterFace
 import com.pbt.myfarm.TaskObject
 import com.pbt.myfarm.Util.AppUtils
-import com.pbt.myfarm.Util.MySharedPreference
 import retrofit2.Call
 import retrofit2.Callback
 import retrofit2.Response
@@ -32,13 +33,35 @@ class ViewTaskObjectViewModel (val activity: Application) : AndroidViewModel(act
     var deleteHttpResponse = MutableLiveData<HttpResponse>()
     var deleteID = MutableLiveData<String>()
     var selectedTaskFUnctionID = MutableLiveData<String>("")
+    var checkStatusobject=MutableLiveData<HttpResponse>()
 
     init {
         taskObjectList = MutableLiveData<List<TaskObject>>()
+        checkStatusobject=MutableLiveData<HttpResponse>()
     }
 
     fun getObjectsList(taskID :String) {
         ApiClient.client.create(ApiInterFace::class.java).getTaskObject(taskID).enqueue(this)
+    }
+    fun checkStatus(taskID: String){
+
+        val apiInterface = ApiClient.client.create(ApiInterFace::class.java)
+
+        val call: Call<HttpResponse> = apiInterface.checkTaskStatus(taskID)
+        call.enqueue(object :Callback<HttpResponse>{
+            override fun onResponse(call: Call<HttpResponse>, response: Response<HttpResponse>) {
+                if (response.isSuccessful && response.body()?.error==false){
+                    val baseHttpResponse=Gson().fromJson<HttpResponse>(Gson().toJson(response.body()),HttpResponse::class.java)
+                    checkStatusobject.postValue(baseHttpResponse)
+                }
+            }
+
+            override fun onFailure(call: Call<HttpResponse>, t: Throwable) {
+                println(t.localizedMessage.toString())
+            }
+
+        })
+
     }
 
     fun getTaskFunctionList(taskID :String,userId: String) {
@@ -55,6 +78,7 @@ class ViewTaskObjectViewModel (val activity: Application) : AndroidViewModel(act
                         Gson().fromJson(Gson().toJson(response.body()!!.data), BaseTaskFunction::class.java)
                     if (!baseTaskFunction.listTask.isNullOrEmpty()) {
                         Log.d("##1234650"," getTaskFunctionList Response  ==> ")
+                        taskTaskFunctionList.value= emptyList()
                         taskTaskFunctionList.postValue(baseTaskFunction.listTask)
                     }
                 }
@@ -65,27 +89,41 @@ class ViewTaskObjectViewModel (val activity: Application) : AndroidViewModel(act
         })
     }
 
-    fun deleteTaskObject(taskID :String) {
-         val apiInterface = ApiClient.client.create(ApiInterFace::class.java)
+    fun deleteTaskObject(taskID: String, context: Context) {
+        showAlertDialog(context,taskID)
+    }
 
-        val call: Call<HttpResponse> = apiInterface.deleteTaskObject(taskID)
+    private fun showAlertDialog(context: Context, taskID: String) {
+        AlertDialog.Builder(context)
+            .setTitle("Delete")
+            .setMessage("Are you sure you want to Delete?")
+            .setPositiveButton("Yes",
+                DialogInterface.OnClickListener { dialog, which ->
+                    val apiInterface = ApiClient.client.create(ApiInterFace::class.java)
 
-        call.enqueue(object : Callback<HttpResponse> {
-            override fun onResponse(call: Call<HttpResponse>, response: Response<HttpResponse>) {
-                if(response.isSuccessful && !response.body()!!.error){
-                    deleteID.postValue(taskID)
-                    deleteHttpResponse.postValue(response.body())
-                }
-            }
+                    val call: Call<HttpResponse> = apiInterface.deleteTaskObject(taskID)
 
-            override fun onFailure(call: Call<HttpResponse>, t: Throwable) {
-            }
-        })
+                    call.enqueue(object : Callback<HttpResponse> {
+                        override fun onResponse(call: Call<HttpResponse>, response: Response<HttpResponse>) {
+                            if(response.isSuccessful && !response.body()!!.error){
+                                deleteID.postValue(taskID)
+                                deleteHttpResponse.postValue(response.body())
+                            }
+                        }
+
+                        override fun onFailure(call: Call<HttpResponse>, t: Throwable) {
+                        }
+                    })
+                })
+            .setNegativeButton(android.R.string.no, null)
+            .setIcon(android.R.drawable.ic_delete)
+            .show()
     }
 
     override fun onResponse(call: Call<BaseHttpResponse>, response: Response<BaseHttpResponse>) {
 
         if(response.isSuccessful && response.body()?.data?.task_objects?.isNullOrEmpty() == false) {
+            taskObjectList.postValue(emptyList())
             taskObjectList.postValue(response.body()?.data?.task_objects)
         }else{
             taskObjectList.postValue(emptyList())
@@ -93,6 +131,8 @@ class ViewTaskObjectViewModel (val activity: Application) : AndroidViewModel(act
     }
 
     override fun onFailure(call: Call<BaseHttpResponse>, t: Throwable) {
+        println("Get VierwTAskoBject list response Failure "+t.localizedMessage.toString())
+
     }
